AppWeb_Tattoo_Match-main/requirements.txt
gittblinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
pandas==2.1.1
numpy==1.26.2
python-dotenv==1.0.1
sendgrid==6.11.0
openpyxl==3.1.5